Can't fall through world by looking down

+++ ObjParser/ppm.h
@@ -1,7 +1,7 @@
 #include <stdbool.h>
 extern bool readPpmImage(
 	char *filename, // name of the file
-	GLfloat *target, // pointer first element of 3D array of float which contains
+	GLubyte *target, // pointer first element of 3D array of float which contains
 					 // the pixels in RGB representation
 	int x, // the x-location in the target array for the lower-left pixel
 	int y, // the location in the target array for the lower-left pixel
